Refactored code.

+++ src/main/scala/com/eleks/socialnetworkanalyser/entities/UserStats.scala
@@ -2,5 +2,5 @@ package com.eleks.socialnetworkanalyser.entities
 
 case class UserStats (name : String,
                       surname : String,
-                      postCount : Int
+                      postCount : Long
                      )
